Added transfer percentage from config to get account balances repo call

+++ src/SFA.DAS.EAS.Employer_Financial.Database/StoredProcedures/GetAccountBalance_ByAccountIds.sql
@@ -1,6 +1,6 @@
 ﻿CREATE PROCEDURE [employer_financial].[GetAccountBalance_ByAccountIds]
 	@AccountIds [employer_financial].[AccountIds] Readonly,
-	@allowancePercentage FLOAT = 0
+	@allowancePercentage DECIMAL(8,5) = 0
 AS
 	SELECT
 		acc.AccountId,
